+++ packages/ohm-js/src/pexprs-assertAllApplicationsAreValid.js
@@ -91,7 +91,7 @@ pexprs.Apply.prototype._assertAllApplicationsAreValid = function(
   }
 
   const isBuiltInApplySyntactic =
-    BuiltInRules && ruleInfo === BuiltInRules.rules.experimentalApplySyntactic;
+    BuiltInRules && ruleInfo === BuiltInRules.rules.applySyntactic;
   const isBuiltInCaseInsensitive =
     BuiltInRules && ruleInfo === BuiltInRules.rules.caseInsensitive;
 
@@ -108,7 +108,7 @@ pexprs.Apply.prototype._assertAllApplicationsAreValid = function(
       throw errors.incorrectArgumentType('a syntactic rule application', arg);
     }
     if (!isSyntactic(arg.ruleName)) {
-      throw errors.experimentalApplySyntacticWithLexicalRuleApplication(arg);
+      throw errors.applySyntacticWithLexicalRuleApplication(arg);
     }
     if (isContextSyntactic) {
       throw errors.unnecessaryExperimentalApplySyntactic(this);
@@ -116,7 +116,7 @@ pexprs.Apply.prototype._assertAllApplicationsAreValid = function(
   }
 
   // ...and that all of the argument expressions only have valid applications and have arity 1.
-  // If `this` is an application of the built-in experimentalApplySyntactic rule, then its arg is
+  // If `this` is an application of the built-in applySyntactic rule, then its arg is
   // allowed (and expected) to be a syntactic rule, even if we're in a lexical context.
   this.args.forEach(arg => {
     arg._assertAllApplicationsAreValid(ruleName, grammar, isBuiltInApplySyntactic);